+++ lib/pow/phoenix/template.ex
@@ -17,12 +17,32 @@ defmodule Pow.Phoenix.Template do
   @doc false
   defmacro __using__(_opts) do
     quote do
-      @after_compile {unquote(__MODULE__), :__after_compile_phoenix_view__}
-
       import unquote(__MODULE__)
 
-      import Pow.Phoenix.HTML.ErrorHelpers, only: [error_tag: 2]
-      import Phoenix.HTML.{Form, Link}
+      unquote do
+        if Pow.dependency_vsn_match?(:phoenix, ">= 1.7.0") do
+          quote do
+            use Phoenix.Component
+            import Pow.Phoenix.HTML.CoreComponents
+            alias Phoenix.LiveView.JS
+          end
+        else
+          # TODO: Remove when Phoenix 1.7 is required
+          quote do
+            import Pow.Phoenix.HTML.ErrorHelpers, only: [error_tag: 2]
+            import Phoenix.HTML.{Form, Link}
+          end
+        end
+      end
+
+      # TODO: Remove when Phoenix 1.7 is required
+      unquote do
+        if Code.ensure_loaded?(Phoenix.View) do
+          quote do
+            @after_compile {unquote(__MODULE__), :__after_compile_phoenix_view__}
+          end
+        end
+      end
 
       unquote do
         # TODO: Remove when Phoenix 1.7 is required
@@ -67,12 +87,53 @@ defmodule Pow.Phoenix.Template do
   """
   @spec template(atom(), atom(), binary() | {atom(), any()}) :: Macro.t()
   defmacro template(action, :html, content) do
-    content = EEx.eval_string(content)
-    render_html_template(action, content)
-  end
+    import_functions = [{__MODULE__, [__inline_route__: 2, __user_id_field__: 2]}]
 
-  defp render_html_template(action, content) do
-    quoted = EEx.compile_string(content, engine: Phoenix.HTML.Engine, line: 1, trim: true)
+    import_functions =
+      if Pow.dependency_vsn_match?(:phoenix, ">= 1.7.0") do
+        import_functions
+      else
+        # TODO: Remove when Phoenix 1.7 required
+        import_functions ++ [{Pow.Phoenix.HTML.FormTemplate, [render_form: 1, render_form: 2]}]
+      end
+
+    content =
+      # TODO: Remove when Phoenix 1.7 required and fallback templates removed
+      if Pow.dependency_vsn_match?(:phoenix, ">= 1.7.0") do
+        String.replace(content, "<%= render_form", "<%% render_form")
+      else
+        content
+      end
+
+    expr =
+      EEx.eval_string(
+        content,
+        [],
+        functions: import_functions,
+        file: __CALLER__.file,
+        line: __CALLER__.line + 1,
+        caller: __CALLER__)
+
+    opts =
+      if Pow.dependency_vsn_match?(:phoenix, ">= 1.7.0") do
+        [
+          engine: Phoenix.LiveView.TagEngine,
+          file: __CALLER__.file,
+          line: __CALLER__.line + 1,
+          caller: __CALLER__,
+          source: expr,
+          tag_handler: Phoenix.LiveView.HTMLEngine]
+      else
+        # TODO: Remove when Phoenix 1.7 required
+        [
+          engine: Phoenix.HTML.Engine,
+          file: __CALLER__.file,
+          line: __CALLER__.line + 1,
+          caller: __CALLER__,
+          source: expr]
+      end
+
+    quoted = EEx.compile_string(expr, opts)
 
     quote do
       def unquote(action)(var!(assigns)) do
@@ -80,10 +141,11 @@ defmodule Pow.Phoenix.Template do
         unquote(quoted)
       end
 
-      def html(unquote(action)), do: unquote(content)
+      def html(unquote(action)), do: unquote(expr)
     end
   end
 
+  @doc false
   if Code.ensure_loaded?(Phoenix.VerifiedRoutes) do
     def __inline_route__(plug, plug_opts) do
       "Pow.Phoenix.Routes.path_for(@conn, #{inspect plug}, #{inspect plug_opts})"
@@ -93,5 +155,16 @@ defmodule Pow.Phoenix.Template do
     def __inline_route__(plug, plug_opts) do
       "Routes.#{Pow.Phoenix.Controller.route_helper(plug)}_path(@conn, #{inspect plug_opts}) %>"
     end
+  end
+
+  @doc false
+  def __user_id_field__(type, :key) do
+    "Pow.Ecto.Schema.user_id_field(#{type})"
+  end
+  def __user_id_field__(type, :type) do
+    "Pow.Ecto.Schema.user_id_field(#{type}) == :email && \"email\" || \"text\""
+  end
+  def __user_id_field__(type, :label) do
+    "Phoenix.Naming.humanize(Pow.Ecto.Schema.user_id_field(#{type}))"
   end
 end