+++ test/pow/phoenix/controllers/session_controller_test.exs
@@ -2,6 +2,7 @@ defmodule Pow.Phoenix.SessionControllerTest do
   use Pow.Test.Phoenix.ConnCase
 
   alias Plug.Conn
+  alias Phoenix.LiveViewTest.DOM
   alias Pow.Plug
 
   describe "new/2" do
@@ -22,11 +23,24 @@ defmodule Pow.Phoenix.SessionControllerTest do
       assert html = html_response(conn, 200)
       assert html =~ Routes.pow_session_path(conn, :create)
       refute html =~ "request_path="
-      assert html =~ "<label for=\"user_email\">Email</label>"
-      assert html =~ "<input id=\"user_email\" name=\"user[email]\" type=\"text\">"
-      assert html =~ "<label for=\"user_password\">Password</label>"
-      assert html =~ "<input id=\"user_password\" name=\"user[password]\" type=\"password\">"
-      assert html =~ "<a href=\"/registration/new\">Register</a>"
+
+      html_tree = DOM.parse(html)
+
+      assert [label_elem] = DOM.all(html_tree, "label[for=user_email]")
+      assert [input_elem] = DOM.all(html_tree, "input[name=\"user[email]\"]")
+      assert DOM.to_text(label_elem) =~ "Email"
+      assert DOM.attribute(input_elem, "type") == "email"
+      refute DOM.attribute(input_elem, "value")
+      assert DOM.attribute(input_elem, "required")
+
+      assert [label_elem] = DOM.all(html_tree, "label[for=user_password]")
+      assert [input_elem] = DOM.all(html_tree, "input[name=\"user[password]\"]")
+      assert DOM.to_text(label_elem) =~ "Password"
+      assert DOM.attribute(input_elem, "type") == "password"
+      refute DOM.attribute(input_elem, "value")
+      assert DOM.attribute(input_elem, "required")
+
+      assert [_] = DOM.all(html, "a[href=\"/registration/new\"]")
     end
 
     test "with request_path", %{conn: conn} do
@@ -43,14 +57,19 @@ defmodule Pow.Phoenix.SessionControllerTest do
         |> get(Routes.pow_session_path(conn, :new))
 
       assert html = html_response(conn, 200)
-      assert html =~ "<label for=\"user_username\">Username</label>"
-      assert html =~ "<input id=\"user_username\" name=\"user[username]\" type=\"text\">"
+
+      html_tree = DOM.parse(html)
+
+      assert [label_elem] = DOM.all(html_tree, "label[for=user_username]")
+      assert [input_elem] = DOM.all(html_tree, "input[name=\"user[username]\"]")
+      assert DOM.to_text(label_elem) =~ "Username"
+      assert DOM.attribute(input_elem, "type") == "text"
     end
   end
 
   describe "create/2" do
-    @valid_params %{"user" => %{"email" => "test@example.com", "password" => "secret"}}
-    @invalid_params %{"user" => %{"email" => "test@example.com", "password" => "invalid"}}
+    @valid_params %{"user" => %{"email" => "mock@example.com", "password" => "secret"}}
+    @invalid_params %{"user" => %{"email" => "invalid@example.com", "password" => "invalid"}}
 
     test "already signed in", %{conn: conn} do
       conn =
@@ -75,8 +94,15 @@ defmodule Pow.Phoenix.SessionControllerTest do
 
       assert html = html_response(conn, 200)
       assert get_flash(conn, :error) == "The provided login details did not work. Please verify your credentials, and try again."
-      assert html =~ "<input id=\"user_email\" name=\"user[email]\" type=\"text\" value=\"test@example.com\">"
-      assert html =~ "<input id=\"user_password\" name=\"user[password]\" type=\"password\">"
+
+      html_tree = DOM.parse(html)
+
+      assert [input_elem] = DOM.all(html_tree, "input[name=\"user[email]\"]")
+      assert DOM.attribute(input_elem, "value") == "invalid@example.com"
+
+      assert [input_elem] = DOM.all(html_tree, "input[name=\"user[password]\"]")
+      refute DOM.attribute(input_elem, "value")
+
       refute Plug.current_user(conn)
       refute conn.private[:plug_session]["auth"]
       refute html =~ "request_path"
@@ -108,11 +134,7 @@ defmodule Pow.Phoenix.SessionControllerTest do
     end
 
     test "removes authenticated", %{conn: conn} do
-      conn = post(conn, Routes.pow_session_path(conn, :create, @valid_params))
-
-      assert %{id: 1} = Plug.current_user(conn)
-      assert conn.private[:plug_session]["auth"]
-      assert_receive {:ets, :put, [{_key, _value} | _rest], _opts}
+      conn = authenticated_conn(conn)
 
       conn = delete(conn, Routes.pow_session_path(conn, :delete))
       assert redirected_to(conn) == "/signed_out"
@@ -120,5 +142,17 @@ defmodule Pow.Phoenix.SessionControllerTest do
       refute Plug.current_user(conn)
       refute conn.private[:plug_session]["auth"]
     end
+  end
+
+  @auth_params %{"user" => %{"email" => "mock@example.com", "password" => "secret"}}
+
+  defp authenticated_conn(conn) do
+    conn = post(conn, Routes.pow_session_path(conn, :create, @auth_params))
+
+    assert %{id: 1} = Plug.current_user(conn)
+    assert conn.private[:plug_session]["auth"]
+    assert_receive {:ets, :put, [{_key, _value} | _rest], _opts}
+
+    conn
   end
 end